+++ Cargo.toml
@@ -24,8 +24,10 @@ members = [
     "core-primitives/storage-verified",
     "core-primitives/teerex-storage",
     "core-primitives/test",
+    "core-primitives/time-utils",
     "core-primitives/types",
     "service",
+    "sidechain/block-composer",
     "sidechain/consensus/aura",
     "sidechain/consensus/common",
     "sidechain/consensus/slots",
@@ -34,6 +36,7 @@ members = [
     "sidechain/sidechain-crate",
     "sidechain/state",
     "sidechain/top-pool",
+    "sidechain/top-pool-executor",
     "sidechain/top-pool-rpc-author",
     "sidechain/validateer-fetch",
 ]
