+++ core-primitives/nonce-cache/Cargo.toml
@@ -20,7 +20,7 @@ sgx = [
 
 [dependencies]
 # sgx dependencies
-sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
+sgx_tstd = { branch = "v1.1.4-testing", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
 
 # local dependencies
