+++ core-primitives/sgx/crypto/src/lib.rs
@@ -16,14 +16,15 @@ pub mod sgx_reexport_prelude {
 	pub use serde_sgx as serde;
 }
 
+pub mod error;
+pub mod traits;
+
 #[cfg(feature = "sgx")]
 pub mod aes;
 #[cfg(feature = "sgx")]
 pub mod ed25519;
-pub mod error;
 #[cfg(feature = "sgx")]
 pub mod rsa3072;
-pub mod traits;
 
 #[cfg(feature = "sgx")]
 pub use self::aes::*;
