+++ enclave-runtime/Cargo.toml
@@ -85,6 +85,7 @@ itc-direct-rpc-server = { path = "../core/direct-rpc-server", default-features =
 itp-extrinsics-factory = { path = "../core-primitives/extrinsics-factory", default-features = false, features = ["sgx"]}
 itp-ocall-api = { path = "../core-primitives/ocall-api", default-features = false }
 itp-nonce-cache = { path = "../core-primitives/nonce-cache", default-features = false, features = ["sgx"] }
+itp-time-utils = { path = "../core-primitives/time-utils", default-features = false, features = ["sgx"] }
 itp-settings = { path = "../core-primitives/settings" }
 itp-sgx-io = { path = "../core-primitives/sgx/io", default-features = false, features = ["sgx"] }
 itp-storage = { path = "../core-primitives/storage", default-features = false, features = ["sgx"] }
