+++ sidechain/consensus/slots/Cargo.toml
@@ -22,6 +22,7 @@ sp-runtime = { version = "4.0.0-dev", default-features = false, git = "https://g
 its-primitives = { path = "../../primitives", default-features = false }
 its-consensus-common = { path = "../common", default-features = false }
 itp-sgx-io = { path = "../../../core-primitives/sgx/io", default-features = false }
+itp-time-utils = { path = "../../../core-primitives/time-utils", default-features = false }
 itp-types = { path = "../../../core-primitives/types", default-features = false }
 
 [dev-dependencies]
@@ -38,10 +39,12 @@ std = [
     "sp-runtime/std",
     "its-primitives/std",
     "its-consensus-common/std",
+    "itp-time-utils/std",
     "itp-sgx-io/std",
 ]
 sgx = [
     "sgx_tstd",
+    "itp-time-utils/sgx",
     "itp-sgx-io/sgx",
     "its-consensus-common/sgx",
 ]
