+++ sidechain/sidechain-crate/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [features]
 default = ["std"]
 std = [
+    "its-block-composer/std",
     "its-consensus-aura/std",
     "its-consensus-common/std",
     "its-consensus-slots/std",
@@ -14,21 +15,25 @@ std = [
     "its-rpc-handler/std",
     "its-state/std",
     "its-top-pool/std",
+    "its-top-pool-executor/std",
     "its-top-pool-rpc-author/std",
     "its-validateer-fetch/std",
 ]
 sgx = [
+    "its-block-composer/sgx",
     "its-consensus-aura/sgx",
     "its-consensus-common/sgx",
     "its-consensus-slots/sgx",
     "its-rpc-handler/sgx",
     "its-state/sgx",
     "its-top-pool/sgx",
+    "its-top-pool-executor/sgx",
     "its-top-pool-rpc-author/sgx",
 ]
 test = ["its-top-pool-rpc-author/test"]
 
 [dependencies]
+its-block-composer = { path = "../block-composer", default-features = false }
 its-consensus-aura = { path = "../consensus/aura", default-features = false }
 its-consensus-common = { path = "../consensus/common", default-features = false }
 its-consensus-slots = { path = "../consensus/slots", default-features = false }
@@ -36,5 +41,6 @@ its-primitives = { path = "../primitives", default-features = false }
 its-rpc-handler = { path = "../rpc-handler", default-features = false }
 its-state = { path = "../state", default-features = false }
 its-top-pool = { path = "../top-pool", default-features = false }
+its-top-pool-executor = { path = "../top-pool-executor", default-features = false }
 its-top-pool-rpc-author = { path = "../top-pool-rpc-author", default-features = false }
 its-validateer-fetch = { path = "../validateer-fetch", default-features = false }
