+++ sidechain/top-pool-executor/Cargo.toml
@@ -1,0 +1,62 @@
+[package]
+name = "its-top-pool-executor"
+version = "0.8.0"
+authors = ["Integritee AG <hello@integritee.network>"]
+edition = "2018"
+resolver = "2"
+
+[dependencies]
+# sgx dependencies
+sgx_types = { branch = "v1.1.4-testing", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+sgx_tstd = { branch = "v1.1.4-testing", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
+
+# local dependencies
+ita-stf = { path = "../../app-libs/stf", default-features = false }
+itp-stf-executor = { path = "../../core-primitives/stf-executor", default-features = false }
+itp-test = { path = "../../core-primitives/test", default-features = false, optional = true }
+itp-time-utils = { path = "../../core-primitives/time-utils", default-features = false }
+itp-types = { path = "../../core-primitives/types", default-features = false }
+its-primitives = { path = "../primitives", default-features = false }
+its-state = { path = "../state", default-features = false }
+its-top-pool-rpc-author = { path = "../top-pool-rpc-author", default-features = false }
+
+# integritee dependencies
+sgx-externalities = { default-features = false, git = "https://github.com/integritee-network/sgx-runtime", branch = "master" }
+
+# sgx enabled external libraries
+thiserror_sgx = { package = "thiserror", git = "https://github.com/mesalock-linux/thiserror-sgx", tag = "sgx_1.1.3", optional = true }
+
+# std compatible external libraries (make sure these versions match with the sgx-enabled ones above)
+thiserror = { version = "1.0", optional = true }
+
+# no-std compatible libraries
+codec  = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
+log = { version = "0.4", default-features = false }
+sp-core = { version = "4.0.0-dev", default-features = false, features = ["full_crypto"], git = "https://github.com/paritytech/substrate.git", branch = "master" }
+sp-runtime = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master"}
+
+
+[features]
+default = ["std"]
+std = [
+    "ita-stf/std",
+    "itp-stf-executor/std",
+    "itp-time-utils/std",
+    "itp-types/std",
+    "its-primitives/std",
+    "its-state/std",
+    "its-top-pool-rpc-author/std",
+    "log/std",
+    "sgx-externalities/std",
+    "thiserror",
+]
+sgx = [
+    "sgx_tstd",
+    "ita-stf/sgx",
+    "itp-stf-executor/sgx",
+    "itp-time-utils/sgx",
+    "itp-types/sgx",
+    "its-state/sgx",
+    "its-top-pool-rpc-author/sgx",
+    "thiserror_sgx",
+]
